model-server/object_segmentation/deploy/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anthropic==0.34.0
anyio==4.4.0
astunparse==1.6.3
attrs==24.2.0
boto3==1.34.162
botocore==1.34.162
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.112.1
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.6.0
google-api-core==2.19.1
google-auth==2.33.0
google-cloud-vision==3.7.4
google-pasta==0.2.0
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.65.4
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
jax==0.4.31
jaxlib==0.4.31
jiter==0.5.0
jmespath==1.0.1
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
# mediapipe==0.10.14
ml-dtypes==0.4.0
namex==0.0.8
numpy==1.26.4
openai==1.40.8
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
# protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
scipy==1.14.0
setuptools==73.0.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.0
starlette==0.38.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
# tensorflow==2.17.0
termcolor==2.4.0
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
